// src/bot/enums/commands.enum.ts
export enum Commands {
  SPAM = "spam",
  HIATUS = "hiatus",
  UNHIATUS = "unhiatus",
  HIATUSME = "hiatusme",
  UNHIATUSME = "unhiatusme",
  COUNTMYMESSAGES = "contarmismensajes",
  SETPOINTSTEMPLATE = "plantillapuntos",
  ADDPOINTS = "sumarpuntos",
  KICK = "kick",
  KISS = "kiss",
  HUG = "hug",
  HELP = "help",
}